requirements.txt
asn1crypto==1.5.1
azure-common==1.1.28
azure-core==1.26.3
azure-functions==1.13.3
azure-storage-blob==12.15.0
azure-storage-common==2.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==40.0.1
filelock==3.10.7
idna==3.4
isodate==0.6.1
mysql-connector-python==8.0.32
oscrypto==1.3.0
packaging==23.0
protobuf==3.20.3
pycparser==2.21
pycryptodomex==3.17
PyJWT==2.6.0
pyOpenSSL==23.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
six==1.16.0
snowflake-connector-python==3.0.2
typing==3.7.4.3
typing_extensions==4.5.0
urllib3==1.26.15
